new qa module and rename bin/fastspecfit-qa to bin/fastqa
